refactor: move props logic to canvas
Moved canvas props logic from react-sdk to builder.
Added new context to pass props store, not global because depends on
params.
