+++ packages/react-sdk/src/props.ts
@@ -1,8 +1,4 @@
-import { useContext, useMemo } from "react";
-import { computed } from "nanostores";
-import { useStore } from "@nanostores/react";
 import type { Instance, Page, Prop, Props, Assets } from "@webstudio-is/sdk";
-import { ReactSdkContext } from "./context";
 import { idAttribute, indexAttribute } from "./tree/webstudio-component";
 
 export type PropsByInstanceId = Map<Instance["id"], Prop[]>;
@@ -93,55 +89,6 @@ export const getPropsByInstanceId = (props: Props) => {
 
 // this utility is be used only for preview with static props
 // so there is no need to use computed to optimize rerenders
-export const useInstanceProps = (instanceId: Instance["id"]) => {
-  const {
-    propsByInstanceIdStore,
-    dataSourcesLogicStore,
-    indexesWithinAncestors,
-  } = useContext(ReactSdkContext);
-  const index = indexesWithinAncestors.get(instanceId);
-  const instancePropsObjectStore = useMemo(() => {
-    return computed(
-      [propsByInstanceIdStore, dataSourcesLogicStore],
-      (propsByInstanceId, dataSourcesLogic) => {
-        const instancePropsObject: Record<Prop["name"], unknown> = {};
-        if (index !== undefined) {
-          instancePropsObject[indexAttribute] = index.toString();
-        }
-        const instanceProps = propsByInstanceId.get(instanceId);
-        if (instanceProps === undefined) {
-          return instancePropsObject;
-        }
-        for (const prop of instanceProps) {
-          // asset and page are normalized to string
-          if (prop.type === "asset" || prop.type === "page") {
-            continue;
-          }
-          if (prop.type === "dataSource") {
-            const dataSourceId = prop.value;
-            const value = dataSourcesLogic.get(dataSourceId);
-            if (value !== undefined) {
-              instancePropsObject[prop.name] = value;
-            }
-            continue;
-          }
-          if (prop.type === "action") {
-            const action = dataSourcesLogic.get(prop.id);
-            if (typeof action === "function") {
-              instancePropsObject[prop.name] = action;
-            }
-            continue;
-          }
-          instancePropsObject[prop.name] = prop.value;
-        }
-        return instancePropsObject;
-      }
-    );
-  }, [propsByInstanceIdStore, dataSourcesLogicStore, instanceId, index]);
-  const instancePropsObject = useStore(instancePropsObjectStore);
-  return instancePropsObject;
-};
-
 export const getInstanceIdFromComponentProps = (
   props: Record<string, unknown>
 ) => {
